refactor: rename AnalyticsAppSwift project to SwiftExample
