+++ Examples/SwiftExample/SwiftExample/Assets.xcassets/AccentColor.colorset/Contents.json
