+++ Examples/SwiftExample/SwiftExample/Assets.xcassets/AppIcon.appiconset/Contents.json
