+++ Examples/SwiftExample/SwiftExample/Assets.xcassets/Contents.json
