+++ Examples/SwiftExample/SwiftExample/ViewControllers/BaseViewController.swift
@@ -1,6 +1,6 @@
 //
 //  BaseViewController.swift
-//  AnalyticsAppSwift
+//  SwiftExample
 //
 //  Created by Satheesh Kannan on 19/04/25.
 //
