+++ Examples/SwiftExample/SwiftExampleTests/SwiftExampleTests.swift
@@ -1,13 +1,13 @@
 //
-//  AnalyticsAppSwiftTests.swift
-//  AnalyticsAppSwiftTests
+//  SwiftExampleTests.swift
+//  SwiftExampleTests
 //
 //  Created by Satheesh Kannan on 30/04/25.
 //
 
 import Testing
 
-struct AnalyticsAppSwiftTests {
+struct SwiftExampleTests {
 
     @Test func example() async throws {
         // Write your test here and use APIs like `#expect(...)` to check expected conditions.
