+++ RudderStackAnalytics.xcworkspace/contents.xcworkspacedata
@@ -5,9 +5,6 @@
       location = "group:RudderStackAnalytics.xcodeproj">
    </FileRef>
    <FileRef
-      location = "group:Examples/AnalyticsAppSwift/AnalyticsAppSwift.xcodeproj">
-   </FileRef>
-   <FileRef
       location = "group:Examples/AnalyticsAppObjC/AnalyticsAppObjC.xcodeproj">
    </FileRef>
    <FileRef
@@ -21,5 +18,8 @@
    </FileRef>
    <FileRef
       location = "group:Examples/SwiftUIExample/SwiftUIExample.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "group:Examples/SwiftExample/SwiftExample.xcodeproj">
    </FileRef>
 </Workspace>
